Exit with error when we fail to create a tunnel
It works well for `-t ppp`, but unfortunately the message for `-t ssh-3`
is strange. The `-oExitOnForwardFailure=yes` option combined with `-f`
for `ssh` does not seem to work as advertised, i.e. that the fork is
supposed to happen only after the tunnel has been successfully
established. Instead, we `sleep` a second and fail at finding the local
device.

Fixes #10