Inherit DNS from SSH-HOST with the -D option
With -D or --inherit-dns, the /etc/resolv.conf will be replaced with the
one from the server. When disconnecting the tunnel, the original one
will be restored. Obviously strange things will happen if you connect
and disconnect several tunnels with the -D option, but there's not too
much to do about that. (resolvconf / openresolv?)

Fixes #6